+++ services/frontend-service/src/index.tsx
@@ -17,7 +17,7 @@ Copyright 2021 freiheit.com*/
 import './assets/index.scss';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Routes } from './v2-ui/Routes';
+import { Routes } from './ui/Routes';
 import { BrowserRouter } from 'react-router-dom';
 
 ReactDOM.render(
