+++ services/frontend-service/src/ui/App/App.test.tsx
